Refactor trajectory and map data classes
This update includes significant refactoring of the trajectory and map data classes. The trajectory psi value was reset, the attitude errors enabled, and the coordinate interpolations and calculation in the Map and related classes were adjusted for more efficiency. The Euler angles calculation was also made more precise in the estimators class.